✨ Flink-State状态编程

+++ flink-learning-atguigu/src/main/java/com/yankee/day05/Flink12_Process_WaterSensorVcInrc_ByState.java
@@ -4,6 +4,9 @@ import com.yankee.bean.WaterSensor_Java;
 import org.apache.flink.api.common.eventtime.SerializableTimestampAssigner;
 import org.apache.flink.api.common.eventtime.WatermarkStrategy;
 import org.apache.flink.api.common.functions.MapFunction;
+import org.apache.flink.api.common.state.ValueState;
+import org.apache.flink.api.common.state.ValueStateDescriptor;
+import org.apache.flink.configuration.Configuration;
 import org.apache.flink.streaming.api.datastream.SingleOutputStreamOperator;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.api.functions.KeyedProcessFunction;
@@ -20,8 +23,8 @@ import java.time.Duration;
  * @description TODO
  * @date 2021/12/27 9:39
  */
-public class Flink10_Process_WaterSensorVcInrc {
-    private static final Logger LOG = LoggerFactory.getLogger(Flink10_Process_WaterSensorVcInrc.class);
+public class Flink12_Process_WaterSensorVcInrc_ByState {
+    private static final Logger LOG = LoggerFactory.getLogger(Flink12_Process_WaterSensorVcInrc_ByState.class);
 
     public static void main(String[] args) throws Exception {
         // 1.获取流执行环境
@@ -49,32 +52,45 @@ public class Flink10_Process_WaterSensorVcInrc {
         // 4.分组
         SingleOutputStreamOperator<WaterSensor_Java> result = watermarks.keyBy(WaterSensor_Java::getId)
                 .process(new KeyedProcessFunction<String, WaterSensor_Java, WaterSensor_Java>() {
-                    private Integer lastVc = Integer.MIN_VALUE;
-                    private Long timerTs = Long.MIN_VALUE;
+                    private ValueState<Integer> vcState;
+                    private ValueState<Long> tsState;
+
+                    @Override
+                    public void open(Configuration parameters) throws Exception {
+                        // 状态初始化
+                        vcState = getRuntimeContext().getState(new ValueStateDescriptor<Integer>("vc-state", Integer.class));
+                        tsState = getRuntimeContext().getState(new ValueStateDescriptor<Long>("ts-state", Long.class));
+                    }
 
                     @Override
                     public void processElement(WaterSensor_Java value, KeyedProcessFunction<String, WaterSensor_Java, WaterSensor_Java>.Context ctx, Collector<WaterSensor_Java> out) throws Exception {
+                        // 上一次的vc
+                        Integer lastVc = vcState.value();
+                        // 定时器
+                        Long timerTs = tsState.value();
                         // 当前的vc与上一次比较
                         Integer vc = value.getVc();
-                        if (vc > lastVc && timerTs == Long.MIN_VALUE) {
-                            // 获取当前的事件时间
-                            long watermark = ctx.timerService().currentWatermark() + 10000L;
-                            LOG.info("当前的watermark：{}", watermark);
-                            // 注册定时器
-                            ctx.timerService().registerEventTimeTimer(watermark);
+                        if (lastVc != null) {
+                            if (vc >= lastVc && timerTs == null) {
+                                // 获取当前的事件时间
+                                long watermark = ctx.timerService().currentWatermark() + 10000L;
+                                LOG.info("当前的watermark：{}", watermark);
+                                // 注册定时器
+                                ctx.timerService().registerEventTimeTimer(watermark);
 
-                            // 更新上一次定时器的时间戳
-                            timerTs = watermark;
-                        } else if (vc < lastVc) {
-                            // 删除定时器
-                            ctx.timerService().deleteEventTimeTimer(timerTs);
-                            LOG.info("删除的定时器时间是：{}", timerTs);
-                            // 恢复timerTs
-                            timerTs = Long.MIN_VALUE;
+                                // 更新上一次定时器的时间戳
+                                tsState.update(watermark);
+                            } else if (vc < lastVc && timerTs != null) {
+                                // 删除定时器
+                                ctx.timerService().deleteEventTimeTimer(timerTs);
+                                LOG.info("删除的定时器时间是：{}", timerTs);
+                                // 恢复timerTs
+                                tsState.clear();
+                            }
                         }
 
                         // 更新上一次vc的值
-                        lastVc = vc;
+                        vcState.update(vc);
 
                         // 正常数据输出
                         out.collect(value);
@@ -83,8 +99,8 @@ public class Flink10_Process_WaterSensorVcInrc {
                     @Override
                     public void onTimer(long timestamp, KeyedProcessFunction<String, WaterSensor_Java, WaterSensor_Java>.OnTimerContext ctx, Collector<WaterSensor_Java> out) throws Exception {
                         ctx.output(new OutputTag<String>("SideOutput"){}, ctx.getCurrentKey() + "连续10s水位线没有下降！");
-                        // 恢复timerTs
-                        timerTs = Long.MIN_VALUE;
+                        // 清空定时器时间的状态
+                        tsState.clear();
                     }
                 });
